added homepage and create page
